git-svn-id: https://svn.code.sf.net/p/olam-model/code/branches/olam_chem@316 2bc74c4e-f164-4307-89b2-97abcff1156f

+++ omodel/omic_tabs.f90
@@ -88,8 +88,8 @@ use consts_coms, only: r8
 
 implicit none
 
-integer, intent(in) :: mrl
-real,    intent(in) :: dtl
+integer,  intent(in) :: mrl
+real(r8), intent(in) :: dtl
 
 integer  :: itc, k, idnc
 real     :: tc, y
